+++ packages/button/src/vars/index.ts
@@ -4,6 +4,8 @@ export const appearances = keyMirror('primary', 'secondary', 'stroke', 'flat')
 
 export const iconAligns = keyMirror('left', 'right')
 
+export const layouts = keyMirror('fullWidth', 'inline')
+
 export const sizes = keyMirror('xSmall', 'small', 'medium', 'large')
 
 export const psds2020Colors = 'psds2020Colors'