+++ packages/core/js/layers.ts
@@ -1,0 +1,7 @@
+export default {
+  main: '0',
+  globalSidenav: '930',
+  globalTopnav: '950',
+  globalDropdown: '1000',
+  globalSkipToContentBanner: '1600'
+}