+++ packages/table/src/css/index.ts
@@ -4,6 +4,7 @@ import {
   colorsBorder,
   colorsTextIcon,
   layout,
+  motion,
   type
 } from '@pluralsight/ps-design-system-core'
 import { names as themeNames } from '@pluralsight/ps-design-system-theme'
@@ -15,7 +16,9 @@ const cellPadding = `${layout.spacingXSmall} ${layout.spacingSmall}`
 const dark = {
   className: `.psds-theme--${themeNames.dark}`,
 
-  borderColor: colorsBorder.lowOnDark,
+  // NOTE: should be using colorsBorder.lowOnDark but border collapse used with
+  //       a transparent color causes weird multi-shade borders
+  borderColor: '#343434',
   rowHoverColor: colorsBackgroundDark[2],
   rowSelectedColor: colorsBackgroundDark[2],
   stickyBgColor: colorsBackgroundDark[1],
@@ -27,7 +30,9 @@ const dark = {
 const light = {
   className: `.psds-theme--${themeNames.light}`,
 
-  borderColor: colorsBorder.lowOnLight,
+  // NOTE: should be using colorsBorder.lowOnLight but border collapse used with
+  //       a transparent color causes weird multi-shade borders
+  borderColor: '#D9D9D9',
   rowHoverColor: colorsBackgroundLight[2],
   rowSelectedColor: colorsBackgroundLight[1],
   stickyBgColor: colorsBackgroundLight[3],
@@ -120,7 +125,7 @@ export default {
     fontWeight: type.fontWeightBook,
     textTransform: 'uppercase',
 
-    '& >:first-of-type': { borderTop: 'none' }
+    '& > tr:first-of-type': { borderTop: 'none' }
   },
 
   [`.psds-table__head${dark.className}`]: { color: dark.textHeadingColor },
@@ -128,11 +133,15 @@ export default {
 
   '.psds-table__header': {
     label: 'table__header',
+    padding: 0,
 
-    fontWeight: type.fontWeightBold,
-    padding: cellPadding,
-    textAlign: 'left',
-    whiteSpace: 'nowrap'
+    '& > div': {
+      fontWeight: type.fontWeightBold,
+      padding: cellPadding,
+      textAlign: 'left',
+      verticalAlign: 'bottom',
+      whiteSpace: 'nowrap'
+    }
   },
   [`.psds-table__header--align-${alignments.center}`]: { textAlign: 'center' },
   [`.psds-table__header--align-${alignments.left}`]: { textAlign: 'left' },
@@ -159,10 +168,25 @@ export default {
 
   '.psds-table__header--sticky': {
     '&[scope="col"]': {
+      border: 'none',
       top: 0,
       position: 'sticky',
-      zIndex: 2
+      zIndex: 2,
+
+      '& > div': { padding: cellPadding }
     },
+
+    // NOTE: used by StickyContainer
+    '&[scope="col"][data-stuck]': {
+      position: 'initial',
+
+      '& > div': {
+        position: 'fixed',
+        top: 0,
+        zIndex: 2
+      }
+    },
+
     '&[scope="row"]': {
       left: 0,
       position: 'sticky',
@@ -171,7 +195,7 @@ export default {
   },
 
   [`.psds-table__header--sticky${dark.className}`]: {
-    '&[scope="col"]': {
+    '&[scope="col"] > div': {
       backgroundColor: dark.stickyBgColor,
       boxShadow: `0 1px 0 0 ${dark.borderColor}`
     },
@@ -181,7 +205,7 @@ export default {
     }
   },
   [`.psds-table__header--sticky${light.className}`]: {
-    '&[scope="col"]': {
+    '&[scope="col"] > div': {
       backgroundColor: light.stickyBgColor,
       boxShadow: `0 1px 0 0 ${light.borderColor}`
     },
@@ -193,21 +217,43 @@ export default {
 
   [`.psds-table__row${dark.className}`]: {
     borderTop: `1px solid ${dark.borderColor}`,
+
     'tbody &:hover > th, tbody &:hover > td': {
       backgroundColor: dark.rowHoverColor
     }
   },
   [`.psds-table__row${light.className}`]: {
     borderTop: `1px solid ${light.borderColor}`,
+
     'tbody &:hover > th, tbody &:hover > td': {
       backgroundColor: light.rowHoverColor
     }
   },
-  '.psds-table__row--collapsed': { display: 'none' },
+  '.psds-table__row--collapsed': {
+    borderColor: 'transparent',
+    height: 0,
+    visibility: 'hidden',
+
+    '& > th, & > td': { paddingBottom: 0, paddingTop: 0 }
+  },
   [`.psds-table__row--selected${dark.className}`]: {
     '& > th, & > td': { backgroundColor: dark.rowSelectedColor }
   },
   [`.psds-table__row--selected${light.className}`]: {
     '& > th, & > td': { backgroundColor: light.rowSelectedColor }
+  },
+
+  '.psds-table__drawer': {
+    borderTopColor: 'transparent',
+
+    'tbody &:hover > th, tbody &:hover > td': { backgroundColor: 'transparent' }
+  },
+  '.psds-table__drawer__cell': {
+    paddingBottom: layout.spacingSmall,
+    paddingTop: layout.spacingSmall,
+    transition: `padding ${motion.speedNormal}`
+  },
+  '.psds-table__drawer__inner': {
+    transition: `all ${motion.speedNormal}`
   }
 }