+++ packages/actionmenu/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [15.4.8](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-actionmenu@15.4.7...@pluralsight/ps-design-system-actionmenu@15.4.8) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-actionmenu
+
+
+
+
+
+## [15.4.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-actionmenu@15.4.6...@pluralsight/ps-design-system-actionmenu@15.4.7) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-actionmenu
+
+
+
+
+
+## [15.4.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-actionmenu@15.4.5...@pluralsight/ps-design-system-actionmenu@15.4.6) (2020-12-16)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-actionmenu
+
+
+
+
+
+## [15.4.5](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-actionmenu@15.4.4...@pluralsight/ps-design-system-actionmenu@15.4.5) (2020-12-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-actionmenu
+
+
+
+
+
 ## [15.4.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-actionmenu@15.4.3...@pluralsight/ps-design-system-actionmenu@15.4.4) (2020-12-11)