+++ packages/linearprogress/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.2.22](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-linearprogress@5.2.21...@pluralsight/ps-design-system-linearprogress@5.2.22) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-linearprogress
+
+
+
+
+
+## [5.2.21](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-linearprogress@5.2.20...@pluralsight/ps-design-system-linearprogress@5.2.21) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-linearprogress
+
+
+
+
+
+## [5.2.20](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-linearprogress@5.2.19...@pluralsight/ps-design-system-linearprogress@5.2.20) (2020-12-16)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-linearprogress
+
+
+
+
+
+## [5.2.19](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-linearprogress@5.2.18...@pluralsight/ps-design-system-linearprogress@5.2.19) (2020-12-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-linearprogress
+
+
+
+
+
 ## [5.2.18](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-linearprogress@5.2.17...@pluralsight/ps-design-system-linearprogress@5.2.18) (2020-12-11)