+++ packages/navcookiebanner/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.12](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navcookiebanner@1.4.11...@pluralsight/ps-design-system-navcookiebanner@1.4.12) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navcookiebanner
+
+
+
+
+
+## [1.4.11](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navcookiebanner@1.4.10...@pluralsight/ps-design-system-navcookiebanner@1.4.11) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navcookiebanner
+
+
+
+
+
+## [1.4.10](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navcookiebanner@1.4.9...@pluralsight/ps-design-system-navcookiebanner@1.4.10) (2020-12-16)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navcookiebanner
+
+
+
+
+
+## [1.4.9](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navcookiebanner@1.4.8...@pluralsight/ps-design-system-navcookiebanner@1.4.9) (2020-12-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navcookiebanner
+
+
+
+
+
 ## [1.4.8](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navcookiebanner@1.4.7...@pluralsight/ps-design-system-navcookiebanner@1.4.8) (2020-12-11)