+++ packages/navitem/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.12](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navitem@1.5.11...@pluralsight/ps-design-system-navitem@1.5.12) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navitem
+
+
+
+
+
+## [1.5.11](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navitem@1.5.10...@pluralsight/ps-design-system-navitem@1.5.11) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navitem
+
+
+
+
+
+## [1.5.10](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navitem@1.5.9...@pluralsight/ps-design-system-navitem@1.5.10) (2020-12-16)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navitem
+
+
+
+
+
+## [1.5.9](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navitem@1.5.8...@pluralsight/ps-design-system-navitem@1.5.9) (2020-12-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navitem
+
+
+
+
+
 ## [1.5.8](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navitem@1.5.7...@pluralsight/ps-design-system-navitem@1.5.8) (2020-12-11)