+++ packages/navuser/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.21](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navuser@1.2.20...@pluralsight/ps-design-system-navuser@1.2.21) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navuser
+
+
+
+
+
+## [1.2.20](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navuser@1.2.19...@pluralsight/ps-design-system-navuser@1.2.20) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navuser
+
+
+
+
+
+## [1.2.19](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navuser@1.2.18...@pluralsight/ps-design-system-navuser@1.2.19) (2020-12-16)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navuser
+
+
+
+
+
+## [1.2.18](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navuser@1.2.17...@pluralsight/ps-design-system-navuser@1.2.18) (2020-12-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navuser
+
+
+
+
+
 ## [1.2.17](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navuser@1.2.16...@pluralsight/ps-design-system-navuser@1.2.17) (2020-12-11)