+++ packages/position/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.19](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-position@5.0.18...@pluralsight/ps-design-system-position@5.0.19) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-position
+
+
+
+
+
+## [5.0.18](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-position@5.0.17...@pluralsight/ps-design-system-position@5.0.18) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-position
+
+
+
+
+
+## [5.0.17](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-position@5.0.16...@pluralsight/ps-design-system-position@5.0.17) (2020-12-16)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-position
+
+
+
+
+
+## [5.0.16](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-position@5.0.15...@pluralsight/ps-design-system-position@5.0.16) (2020-12-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-position
+
+
+
+
+
 ## [5.0.15](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-position@5.0.14...@pluralsight/ps-design-system-position@5.0.15) (2020-12-11)