+++ packages/row/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.0.10](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-row@11.0.9...@pluralsight/ps-design-system-row@11.0.10) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-row
+
+
+
+
+
+## [11.0.9](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-row@11.0.8...@pluralsight/ps-design-system-row@11.0.9) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-row
+
+
+
+
+
+## [11.0.8](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-row@11.0.7...@pluralsight/ps-design-system-row@11.0.8) (2020-12-16)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-row
+
+
+
+
+
+## [11.0.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-row@11.0.6...@pluralsight/ps-design-system-row@11.0.7) (2020-12-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-row
+
+
+
+
+
 ## [11.0.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-row@11.0.5...@pluralsight/ps-design-system-row@11.0.6) (2020-12-11)