+++ packages/scrollable/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-scrollable@3.0.6...@pluralsight/ps-design-system-scrollable@3.0.7) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-scrollable
+
+
+
+
+
+## [3.0.6](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-scrollable@3.0.5...@pluralsight/ps-design-system-scrollable@3.0.6) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-scrollable
+
+
+
+
+
+## [3.0.5](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-scrollable@3.0.4...@pluralsight/ps-design-system-scrollable@3.0.5) (2020-12-16)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-scrollable
+
+
+
+
+
+## [3.0.4](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-scrollable@3.0.3...@pluralsight/ps-design-system-scrollable@3.0.4) (2020-12-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-scrollable
+
+
+
+
+
 ## [3.0.3](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-scrollable@3.0.2...@pluralsight/ps-design-system-scrollable@3.0.3) (2020-12-11)