+++ packages/tooltip/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.2.19](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tooltip@6.2.18...@pluralsight/ps-design-system-tooltip@6.2.19) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-tooltip
+
+
+
+
+
+## [6.2.18](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tooltip@6.2.17...@pluralsight/ps-design-system-tooltip@6.2.18) (2020-12-17)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-tooltip
+
+
+
+
+
+## [6.2.17](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tooltip@6.2.16...@pluralsight/ps-design-system-tooltip@6.2.17) (2020-12-16)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-tooltip
+
+
+
+
+
+## [6.2.16](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tooltip@6.2.15...@pluralsight/ps-design-system-tooltip@6.2.16) (2020-12-14)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-tooltip
+
+
+
+
+
 ## [6.2.15](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tooltip@6.2.14...@pluralsight/ps-design-system-tooltip@6.2.15) (2020-12-11)